improve definitions of structures and unions

+++ src/test/csrc/common/golden.h
@@ -27,24 +27,24 @@ extern "C" uint8_t pte_helper(uint64_t satp, uint64_t vpn, uint64_t *pte, uint8_
 
 typedef union PageTableEntry {
   struct {
-    uint32_t v   : 1;
-    uint32_t r   : 1;
-    uint32_t w   : 1;
-    uint32_t x   : 1;
-    uint32_t u   : 1;
-    uint32_t g   : 1;
-    uint32_t a   : 1;
-    uint32_t d   : 1;
-    uint32_t rsw : 2;
+    uint64_t v   : 1;
+    uint64_t r   : 1;
+    uint64_t w   : 1;
+    uint64_t x   : 1;
+    uint64_t u   : 1;
+    uint64_t g   : 1;
+    uint64_t a   : 1;
+    uint64_t d   : 1;
+    uint64_t rsw : 2;
     uint64_t ppn :44;
-    uint32_t pad :10;
+    uint64_t pad :10;
   };
   struct {
-    uint32_t difftest_v    : 1;
-    uint32_t difftest_perm : 7;
-    uint32_t difftest_rsw  : 2;
+    uint64_t difftest_v    : 1;
+    uint64_t difftest_perm : 7;
+    uint64_t difftest_rsw  : 2;
     uint64_t difftest_ppn  :44;
-    uint32_t difftest_pad  :10;
+    uint64_t difftest_pad  :10;
   };
   uint64_t val;
 } PTE;
@@ -52,8 +52,8 @@ typedef union PageTableEntry {
 typedef union atpStruct{
   struct {
     uint64_t ppn: 44;
-    uint32_t asid: 16;
-    uint32_t mode: 4;
+    uint64_t asid: 16;
+    uint64_t mode: 4;
   };
   uint64_t val;
 } Satp, Hgatp;
